+++ transforms/matmul_pad.mlir
@@ -5,7 +5,7 @@ module attributes {transform.with_named_sequence} {
 
     %padded, %pad, %copy_back = transform.structured.pad %matmuls
       pad_to_multiple_of [8, 8, 8] {
-        padding_values = [0.0 : f32, 0.0 : f32, 0.0 : f32],
+        padding_values = [0 : i8, 0 : i8, 0 : i32],
         padding_dimensions = [0, 1, 2],
         nofold_flags = [1, 1, 0]
       } : (!transform.any_op) -> (!transform.any_op, !transform.any_op, !transform.any_op)
